# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}},"post":{"tags":["country-controller"],"operationId":"postCountry","parameters":[{"name":"amountOfCountryToSave","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/name/{name}":{"get":{"tags":["country-controller"],"operationId":"getByName","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"code","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getByMostBorders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/api/countries/language/{language}":{"get":{"tags":["country-controller"],"operationId":"getByLanguage","parameters":[{"name":"language","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/api/countries/continent/{continent}":{"get":{"tags":["country-controller"],"operationId":"getByContinent","parameters":[{"name":"continent","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}}},"components":{"schemas":{"CountryDTO":{"type":"object","properties":{"name":{"type":"string"},"Code":{"type":"string"}}},"Country":{"type":"object","properties":{"name":{"type":"string"},"population":{"type":"integer","format":"int64"},"area":{"type":"number","format":"double"},"code":{"type":"string"},"region":{"type":"string"},"borders":{"type":"array","items":{"type":"string"}},"languages":{"type":"object","additionalProperties":{"type":"string"}}}}}}}